jobs/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgi-lifespan==2.1.0
attrs==23.2.0
beanie==1.26.0
cachetools==5.3.3
certifi==2024.6.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coolname==2.2.0
croniter==2.0.5
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fsspec==2024.6.0
graphviz==0.20.3
griffe==0.45.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
importlib_resources==6.1.3
itsdangerous==2.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy-model==0.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
motor==3.4.0
openai==1.31.0
orjson==3.10.3
packaging==24.0
pathspec==0.12.1
pendulum==3.0.0
prefect-client==2.19.3
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sniffio==1.3.1
text-unidecode==1.3
time-machine==2.14.1
toml==0.10.2
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
uvicorn==0.28.1
websockets==12.0
